+++ src/app/admin/admin.module.ts
@@ -1,0 +1,18 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { ReactiveFormsModule } from '@angular/forms';
+
+import { SharedModule } from '../shared/shared.module';
+import { AdminComponent } from './admin.component';
+import { AdminRoutingModule } from './admin.routing';
+
+@NgModule({
+  declarations: [AdminComponent, AdminRoutingModule.components],
+  imports: [
+    CommonModule,
+    AdminRoutingModule,
+    SharedModule,
+    ReactiveFormsModule,
+  ],
+})
+export class AdminModule {}