+++ src/app/page-forbidden/components/page-forbidden.component.ts
@@ -1,0 +1,8 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+  selector: 'app-page-forbidden',
+  templateUrl: './page-forbidden.component.html',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class PageForbiddenComponent {}