+++ src/app/products/components/index.ts
@@ -1,0 +1,3 @@
+export * from './product';
+export * from './product-list';
+export * from './product-view';