+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,20 +1,23 @@
-import { CartListService } from 'src/app/cart/services/cart-list.service';
-
 import { Component } from '@angular/core';
+import { CartListService } from 'src/app/cart/services/cart-list.service';
 
 import { IProduct } from '../../../shared/models/product.model';
 import { ProductsService } from '../../services/products.service';
+import { CanActivate } from '@angular/router';
 
 @Component({
-  selector: 'app-product-list',
   templateUrl: './product-list.component.html',
   styleUrls: ['./product-list.component.scss', '../../../app.component.scss'],
 })
-export class ProductListComponent {
+export class ProductListComponent implements CanActivate {
   constructor(
     public productService: ProductsService,
     private cartListService: CartListService
   ) {}
+
+  canActivate(): boolean {
+    return false;
+  }
 
   addProductToCart(product: IProduct): void {
     this.cartListService.addProduct(product);