+++ src/app/products/products.module.ts
@@ -1,12 +1,26 @@
-import { NgModule } from '@angular/core';
+import { inject, InjectionToken, NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
 
 import { SharedModule } from '../shared/shared.module';
-import { ProductListComponent } from './components/product-list/product-list.component';
-import { ProductComponent } from './components/product/product.component';
+import { ProductComponent } from './components';
+import { ProductsRoutingModule } from './products.routing';
+import { GeneratorService } from '../core';
 
 @NgModule({
-  declarations: [ProductComponent, ProductListComponent],
-  exports: [ProductComponent, ProductListComponent],
-  imports: [SharedModule],
+  declarations: [ProductsRoutingModule.components, ProductComponent],
+  imports: [CommonModule, SharedModule, ProductsRoutingModule],
+  exports: [ProductComponent],
+  providers: [GeneratorService],
 })
 export class ProductsModule {}
+
+export const GENERATOR_ID_TOKEN = new InjectionToken<(n: number) => string>(
+  'String Generator',
+  {
+    providedIn: ProductsModule,
+    factory: () => {
+      const generatorService = inject(GeneratorService);
+      return (n: number) => generatorService.generate(n);
+    },
+  }
+);