+++ src/app/shared/models/product.model.ts
@@ -9,6 +9,9 @@ export interface IProduct {
   name: string;
   description: string;
   price: number;
+  quantity: number;
   category: Category;
   isAvailable: boolean;
+  imageUrl: string;
+  imageAlt?: string;
 }